backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandon Lamer-Connolly", email = "brandon.lc.1334@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-gis (>=1.2.0,<2.0.0)",
    "dj-database-url (>=2.1.0,<3.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false